+++ arduino/imd.h
@@ -7,7 +7,7 @@ class IseMotorDriver {
   public:
     IseMotorDriver(int8_t addr) : addr(addr) {};
     void set_power(int8_t power);
-    int64_t get_encoder();
+    int32_t get_encoder();
 
   private:
     int8_t addr;
